requirements.txt
aiosqlite==0.17.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.10.0
Deprecated==1.2.13
future==0.18.2
glob2==0.7
greenlet==1.1.2
grpcio==1.41.1
grpcio-tools==1.41.1
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
pefile==2021.9.3
pika==1.2.0
Pillow==8.4.0
plotext==3.1.3
prettytable==2.2.1
prompt-toolkit==3.0.21
protobuf==3.19.0
pycparser==2.20
pyelftools==0.27
pyfiglet==0.8.post1
PyGithub==1.55
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pypager==3.0.0
pyparsing==3.0.1
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.13.2
pyftpdlib==1.5.6
sqlalchemy-utils==0.38.2
boto3==1.23.4
google-api-python-client==2.54.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
